# Backend/public/swagger.json
{
  "components": {
    "schemas": {
      "LoginWithToken": {
        "properties": {
          "status": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": ["status", "message"],
        "type": "object",
        "additionalProperties": false
      },
      "UserCredentialInput": {
        "properties": {
          "username": {
            "type": "string"
          },

          "password": {
            "type": "string"
          }
        },
        "required": ["username", "password"],
        "type": "object",
        "additionalProperties": false
      },
      "UserStatus": {
        "properties": {
          "username": {
            "type": "string"
          },
          "lastStatus": {
            "type": "string"
          }
        },
        "required": ["username", "lastStatus"],
        "type": "object",
        "additionalProperties": false
      },
      "PublicMessage": {
        "properties": {
          "id": {
            "type": "number"
          },
          "content": {
            "type": "string"
          },

          "time": {
            "type": "string"
          },
          "sender": {
            "type": "string"
          }
        },
        "required": ["content", "time", "sender"],
        "type": "object",
        "additionalProperties": false
      },
      "JwtToken": {
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": ["token"],
        "type": "object",
        "additionalProperties": false
      },
      "lastStatus": {
        "properties": {
          "lastStatus": {
            "type": "string"
          }
        },
        "required": ["lastStatus"],
        "type": "object",
        "additionalProperties": false
      },
      "Message": {
        "properties": {
          "id": {
            "type": "number"
          },
          "content": {
            "type": "string"
          },

          "time": {
            "type": "string"
          },
          "sender": {
            "type": "string"
          },
          "sendee": {
            "type": "string"
          }
        },
        "required": ["content", "time", "sender", "sendee"],
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "esn api",
    "version": "0.0.1",
    "license": {
      "name": "MIT"
    },
    "contact": {
      "name": "sb1"
    }
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/users/register": {
      "post": {
        "operationId": "Register",
        "responses": {
          "200": {
            "description": "LoginWithToken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginWithToken"
                }
              }
            }
          }
        },
        "description": "Register a user using the provided username and password",
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCredentialInput"
              }
            }
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "operationId": "LoginUser",
        "responses": {
          "200": {
            "description": "LoginWithToken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginWithToken"
                }
              }
            }
          }
        },
        "description": "login a user with provided username and password",
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCredentialInput"
              }
            }
          }
        }
      }
    },

    "/api/users/:username/status": {
      "get": {
        "operationId": "getUserStatusByUsername",
        "responses": {
          "200": {
            "description": "user status",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/lastStatus"
                  },
                  "lastStatus": { "type": "string" }
                }
              }
            }
          }
        },
        "description": "get the selected user's status with their username as api",
        "security": [],
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string"
            },
            "name": "username",
            "in": "path"
          }
        ]
      }
    },
    "/api/users/status": {
      "get": {
        "operationId": "getUserStatus",
        "responses": {
          "200": {
            "description": "user status",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserStatus"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "description": "Update Current User's Status",
        "security": [],
        "parameters": []
      }
    },
    "/api/messages": {
      "post": {
        "operationId": "post message",
        "responses": {
          "200": {
            "description": "PostMessage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "description": "post a message",
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Message"
              }
            }
          }
        }
      }
    },
    "/api/messages/:sender/:sendee": {
      "get": {
        "operationId": "get all messages from the sender and sendee",
        "responses": {
          "200": {
            "description": "GetMessages"
          }
        },
        "description": "get all messages from the sender and sendee",
        "security": [],
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string"
            },
            "name": "sender",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "type": "string"
            },
            "name": "sendee",
            "in": "path"
          }
        ]
      }
    },
    "/api/messages/:sendee": {
      "get": {
        "operationId": "get all unread messages for sendee",
        "responses": {
          "200": {
            "description": "GetMessages"
          },
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/Message"
                },
                "type": "array"
              }
            }
          }
        },
        "description": "get all unread messages for sendee",
        "security": [],
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string"
            },
            "name": "sendee",
            "in": "path"
          }
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}